refactor(db,sqlite): split each model-sepecific impl. to separate files
