+++ src/db/mod.rs
@@ -3,7 +3,7 @@ pub mod mongodb;
 pub mod postgres;
 pub mod shared;
 pub mod sqlite;
-pub mod sqlite_constructs;
+pub mod sqlite_internals;
 
 pub use shared::Adapter;
 pub use shared::Client;
