+++ src/db/shared.rs
@@ -1,4 +1,4 @@
-use crate::models::{group, Proposal};
+use crate::models::Proposal;
 
 use super::sqlite::SQLite;
 
@@ -37,18 +37,9 @@ pub trait DbAdapter {
 
     /// Removes the all storage for the configured adapter, no data will be saved.
     fn remove(&self) -> Result<Status, DbError>;
-
-    /// Retrieves a single proposal that belong to given group.
-    fn get_proposal(&self, title: &str, owner: &str) -> Result<Proposal, DbError>;
-
-    /// Fetches all proposals that belong to a given group.
-    fn get_proposals(&self, owner: &str) -> Result<Vec<Proposal>, DbError>;
-
-    fn insert_proposal(&self, p: Proposal) -> Result<Proposal, DbError>;
-
-    fn delete_proposal(&self, p: &Proposal) -> Result<(), DbError>;
 }
 
+/// Defines proposal access pattern for each db-client implementation.
 pub trait ProposalAdapter {
     /// Retrieves a single proposal that belong to given group.
     fn get_proposal(&self, title: &str, owner: &str) -> Result<Proposal, DbError>;
@@ -65,12 +56,10 @@ pub struct Client {}
 
 impl Client {
     /// Creates a new client using given adapter.
-    pub fn new(adapter: Adapter, file: &str) -> Box<dyn DbAdapter> {
-        let client = match adapter {
+    pub fn create(adapter: Adapter, file: &str) -> impl DbAdapter + ProposalAdapter {
+        match adapter {
             Adapter::Sqlite => SQLite::new(file),
             _ => panic!("db adapter is not implemented"),
-        };
-
-        Box::new(client)
+        }
     }
 }
