+++ src/repositories/proposal_repository.rs
@@ -1,7 +1,7 @@
-use crate::{db::Client, models::proposal::Proposal};
+// use crate::{db::Client, models::proposal::Proposal};
 
 pub struct ProposalRepository {
-    client: Client,
+    // client: Client,
 }
 
 impl ProposalRepository {}
